bfloat

+++ train.py
@@ -28,7 +28,7 @@ def train_epoch(model, optimizer, criterion, dataloader, scaler, scheduler=None)
     model.train()
     for i, (padded_seq, lengths) in enumerate(tqdm(dataloader, desc="Train epoch", leave=False)):
         optimizer.zero_grad()
-        with torch.autocast(device_type=amp_device, dtype=torch.float16, enabled=True):
+        with torch.autocast(device_type=amp_device, dtype=torch.bfloat16, enabled=True):
             tokens = padded_seq[:, :lengths.max()].to(device)
             logits = model(tokens[:, :-1])
             loss = criterion(logits.transpose(1, 2), tokens[:, 1:])
@@ -50,7 +50,7 @@ def evaluate_epoch(model, criterion, dataloader, scaler):
 
     model.eval()
     for i, (padded_seq, lengths) in enumerate(tqdm(dataloader, desc="Val epoch", leave=False)):
-        with torch.autocast(device_type=amp_device, dtype=torch.float16, enabled=True):
+        with torch.autocast(device_type=amp_device, dtype=torch.bfloat16, enabled=True):
             tokens = padded_seq[:, :lengths.max()].to(device)
             logits = model(tokens[:, :-1])
             loss = criterion(logits.transpose(1, 2), tokens[:, 1:])
